refactor filters logic making responsable to parent component
