+++ src/components/FiltersMobile/FiltersMobile.stories.tsx
@@ -11,13 +11,7 @@ export default {
 } as ComponentMeta<typeof FiltersMobile>;
 
 const Template: ComponentStory<typeof FiltersMobile> = (args) => {
-    const [filteredProducts, setFilteredProducts] = useState(
-        featuredProducts.results
-    );
-
-    console.log('filteredProducts', filteredProducts);
-
-    return <FiltersMobile {...args} setParentProducts={setFilteredProducts} />;
+    return <FiltersMobile {...args} />;
 };
 
 export const Default = Template.bind({});
