+++ src/helpers/filtersHelpers.tsx
@@ -4,6 +4,8 @@ export const filterProductsByCategory = (
     filtersSelected: string[],
     products: ResultsEntity[]
 ) => {
+    console.log('filtersSelected', filtersSelected);
+    console.log('products drom filter fn', products);
     if (filtersSelected.length === 0) return products;
     const filteredProducts = products.filter((item) =>
         filtersSelected.includes(item.data.category.slug)
